restrucutre menu component
