+++ src/components/Menu/index.jsx
@@ -1,35 +1,3 @@
-import { MenuToggleButton } from "../Button";
-import MenuDesktop from "./MenuDesktop";
-import MenuMobile from "./MenuMobile";
-import DeviceMenu from "./DeviceMenu";
-
-const Menu = ({
-  menuItems,
-  darkMode,
-  switchTheme,
-  isOpen = false,
-  viewportWidth = 769,
-  toggleMenu,
-}) => {
-  const scrollToRef = (ref) => {
-    window.scrollTo({ top: ref.current.offsetTop - 65, behavior: "smooth" });
-    toggleMenu();
-  };
-
-  return (
-    <>
-      <DeviceMenu
-        menuItems={menuItems}
-        darkMode={darkMode}
-        switchTheme={switchTheme}
-        isOpen={isOpen}
-        viewportWidth={viewportWidth}
-        toggleMenu={toggleMenu}
-        scrollToRef={scrollToRef}
-      />
-      {!isOpen && <MenuToggleButton toggleMenu={toggleMenu} />}
-    </>
-  );
-};
+import Menu from "./Menu";
 
 export default Menu;
